Move state related to map from AppState to MapState
