+++ src/App.tsx
@@ -83,8 +83,8 @@ const App = ({ dispatch, isEmbedMode }) => {
   );
 };
 
-const mapStateToProps = (state) => ({
-  isEmbedMode: state.AppState.isEmbedMode
+const mapStateToProps = (state: RootState) => ({
+  isEmbedMode: state.MapState.isEmbedMode
 });
 
 export default hot(connect(mapStateToProps)(App));
